moved msbuild source mapping tests in a new file to take advantage of parallel execution
